refactored the main feature processors

+++ src/main/java/fact/features/PhotonCharge.java
@@ -25,12 +25,12 @@ import fact.Constants;
  */
 public class PhotonCharge implements Processor {
 	static Logger log = LoggerFactory.getLogger(PhotonCharge.class);
-	private float[] photonCharge = null;
+	private double[] photonCharge = null;
 
 	private String color = "#00F0F0";
 
 	private	double average = 0.0;
-	private float integralGain = 244.0f;
+	private double integralGain = 244.0f;
 	private int alpha = 64;
 
 	private String positions = null;
@@ -40,9 +40,9 @@ public class PhotonCharge implements Processor {
 	@Override
 	public Data process(Data input) {
 		int[] posArray = null;
-		float[] data = null;
+		double[] data = null;
 		try{
-			data = (float[]) input.get(key);
+			data = (double[]) input.get(key);
 			if (data == null){
 				throw new Exception();
 			}
@@ -58,7 +58,7 @@ public class PhotonCharge implements Processor {
 		}
 
 		IntervalMarker[] m = new IntervalMarker[Constants.NUMBEROFPIXEL];
-		photonCharge = new float[Constants.NUMBEROFPIXEL];
+		photonCharge = new double[Constants.NUMBEROFPIXEL];
 		int roi = data.length / Constants.NUMBEROFPIXEL;
 		// for each pixel
 		for(int pix = 0 ; pix < Constants.NUMBEROFPIXEL; pix++){
@@ -117,7 +117,7 @@ public class PhotonCharge implements Processor {
 
 	/*Getters and Setters */
 
-	public float getIntegralGain() {
+	public double getIntegralGain() {
 		return integralGain;
 	}
 	@Parameter(required = false, description = "Value for the integral Gain. This is a measured Constant.", defaultValue = "244.0")
